calculate_tf_idf and relevand_doc have some mistakes
at this point i am tired and if i look at anything more it will be on sunday as i have other things to work on between now and sunday. Would like it if someone else could take a look and help out. Will help with readme/recording for sunday.